Deduplicate staging table fields name and not by DDL string
The purpose of this PR is to start deduplicating the `fields_dml/ddl` (`data_vault_load.py`) by field name and not by `DML/DDL` expression.
In theory, the same field have the same `DDL/DML` in the staging table, independently of the `DV` table in which it is stored (*Example:* `article_id` should have the same DDL in `h_article` and `l_order_article`) so we should never have duplicate fields, even if we deduplicate by the `DML/DDL` string. Even though, it is not functionally correct to deduplicate fields by its SQL expression, when what we want to assure is that the same *field name* is not repeated in the staging table.

While there:
- Changed `DataVaultField` class to a regular class (instead of a `dataclass`), per @zeebonk's suggestion. In the first implementation of this class, it made sense to make it a `dataclass`, but then we started to use it as a regular class, applying logic and overriding `dataclass` pre-implemented methods (`__post_init__` created to lower the case of field and table names; `__eq__` and __hash__` overridden).