Duplicate records entered into Data Vault hub when an ID occurs in several source systems
As Data Vault framework allows one staging table to have multiple sources with the same `r_timestamp`, the previous query templates for Links and Hubs (migrated from Pentaho framework without a change) are not applicable anymore.

The previous implementation applied a `DISTINCT` on the whole Link/Hub data set, which was enough as we only had one source per staging table. In the new version we have to apply a more deterministic deduplication approach (as we do for Satellites), grouping the source data set by the target table hashkey.

## Implementation:
- SQL templates updates;
- Python code adjusted to produce new placeholders.