Change validate() methods to use try/except blocks instead if/else in DV framework
The purpose of this PR is to change all validate methods to start using try/catch blocks, instead of if/else. The previous implementation was always executing a get() command and then evaluate the result of that get(), throwing an exception if the result of the get() command was None. I think it makes more sense to catch a KeyError exception, while using the [] notation.