Enable role-playing hubs in the DV configuration
[JIRA]

The purpose of this PR is to extend the Data Vault framework to allow the usage of role playing hubs.

Role playing hubs are views that point to a parent hub (materialized as a table), similar to the role playing dimensions in our DWH (dm_date, for example). The existence of these types of hubs is needed when we need to define more than one role to a specific entity. One example of this is the h_account, as one account may have more than one role (supplier, administrator, handler, transporter, etc).