Fix role playing hubs to ignore parent hub in staging table creation
The purpose of this PR is to assure that `picnic-data-vault` only considers fields that will actually be populated during the data vault load. In the current implementation, `snowflake.py` deserializes every table included in `target_tables`, even if that table was only used to defined the relationship between a role playing hub and its parent table (that will never be directly populated, but indirectly via its child role playing hubs).

The solution to this issue was the following:
1. Separation of concerns between `target_tables` and `role_playing_hubs` parent tables. In the new implementation, a table should only be included in `target_tables` argument if it will actually be directly populated (not indirectly, via its child tables);
2. `_fields` property was changed to fetch metadata from both `target_tables` and `role_playing_hubs` dictionary values (parent_tables);
3. `deserialized_target_tables` updated to deserialize `target_tables` and `role_playing_hubs` parent tables separately. `role_playing_hubs` parent tables are only deserialized to build the relationship with its child tables.
4. Test suite updated to test two different `DataVaultLoad` object (with and without role playing).