Fix satellites loading for multiple sources (with qualify instead of aggregation)
After the latest release of `picnic-data-vault`, a bug was introduced, specifically for satellite loading from multiple sources. In order to allow loading data from multiple sources, using the same `extraction_timestamp`, the same approach used for hubs and links was followed (to aggregate the data by `hashkey` and fetch the `MIN` of all other fields).

This approach is indeed valid for links and hubs, as each `hashkey` uniquely identifies one record, but not for satellites as the descriptive fields might change between sources. During the aggregation using the `MIN`, we might be fetching each descriptive field from a different source, which is not correct at all. In order to solve this issue, the solution was changed to use a `QUALIFY` filter instead of aggregating the data.

# Implementation:
- Removed all `descriptive_fields_aggregation ` from the satellite query placeholders;
- Adapted the template SQL to perform a `QUALIFY` instead of the aggregation.
- Adapted test suite for the new query results.

# While there:
- Added `parent_table=None` command in `Satellite` class as there was no reference to this field anywhere and, even though this field is set after instantiation (not in `__init__`), it's always good to have the reference to the field existence.