Fix rank column name clash in picnic-data-vault
The purpose of this PR is to  rename`rank` field in both satellites queries to `_rank`. The name that is now in place is prone to crashing with actual data vault fields (which just happened during the development of the banners domain).

With `_rank` we can make sure that it never clashes with any of our data vault model (no fields with underscores exist or will ever exist).

## Implementation:
Renamed `rank` field to `_rank`.

## While there:
Bumped the version of `picnic-database`.