Fix picnic-data-vault to evaluate prefixes and suffixes with `==`
The purpose of this PR is to update `picnic-data-vault` to compare a field prefix/suffix against `FIELD_PREFIX`/`FIELD_SUFFIX` dictionaries, using `==` instead of `in`.

`FIELD_PREFIX`/`FIELD_SUFFIX` are `Dict[FieldRole, str]`, and not `Dict[FieldRole, list]`, as we have in `TABLE_PREFIXES` dict. There's only one prefix/suffix per field role, but more than one possible prefix for a table type.

The explanation for this issue is that, at the time of development, all three dictionaries were built at the same time, all as `Dict[FieldRole, list]`. Then I simplified the code to keep `Dict[FieldRole, list]` only in the one that actually needed it (`TABLE_PREFIXES`), but didn't adapt this part of the code to compare using `==` instead of `in`.

As `in` is also a valid function to apply in a `str`, this didn't generate any error/test failure so was never detected.

Now @tsferreira4 needed to use `picnic-data-vault` to populate a field called `grouping_key`, that was identified as a `hashkey_parent`, because `key` exists in `hashkey`!

## Implementation:
- Changed the code to use `==` instead of `in` while attributing the field role;
- Added test case for this edge case.