Change picnic-data-vault to infer field role based on previous (Pentaho rules)
# Purpose:

On the context of this comment: [Github] the `picnic-data-vault` package should be updated to apply the Pentaho rule to infer `FieldRole=Business key`.

# Implementation:

`DataVaultField` role property updated to follow previous logic.

The only difference here is that the rule suggested in the comment (and applied in Pentaho) does not consider the child keys prefix. In this package it is considered.

# Dependencies:
Dependency for: [Github]