Change picnic-data-vault to access metadata table
## Issue:
As the metadata in our Snowflake database is getting bigger and bigger, we started experiencing a known issue from the past, when the queries to `information_schema` randomly fail with `Information schema query returned too much data.  Please repeat query with more selective predicates.`
Even though the error message refers to the limit as being based on the number of rows returned by the query on `information_schema`, during some past chat with the Snowflake support, we were told that this limit is also based on other variables like the number of objects in the database, the number of queries running simultaneous, etc.

## Solution:
Similarly to what was done in Pentaho, the solution is to access our metadata table (refreshed on each DV deployment) instead of the live query on `information_schema`.

## Implementation:
Changed `src/picnic/data_vault/deserializers/snowflake_model_metadata.sql` to access `cfg.dv_metadata` instead of `information_schema.columns` directly.

## Notes:
- This should be a temporary solution, as the correct solution is to use `SHOW COLUMNS IN SCHEMA` instead of querying the information schema directly. `SHOW` commands have no limitation in terms of number of rows, and are compatible with all snowflake databases. 
- I did not change the code to use `SHOW` commands in this PR as we need a quick solution to ensure nobody wakes up during the night! Changing to the `SHOW` commands may seem simple, but the data returned by this command, comparatively to our current query is quite different and would demand more development time.