Update picnic-data-vault to use Snowflake SHOW COLUMNS
The purpose of this PR is to refactor `picnic-data-vault`'s `SnowflakeDeserializer` to start performing a live query to `Snowflake`'s metadata instead of accessing `cfg.dv_metadata` table (introduced in [Github] 

The issue that made us change the code to use `cfg.dv_metadata` table only occurs when querying `information_schema` views, but not using `SHOW COLUMNS` command, as `SHOW` commands do not have any limitation on the number of records fetched.

The change to the `SHOW COLUMNS` command demands a bit more  data transformation on the Python side, as `SHOW` commands are not queryable, so every data transformation that is currently done in `SQL` have to be done in Python. The major differences between the data set returned by `SHOW COLUMNS` and our current metadata query are the following:
1. `SHOW COLUMNS` does not return the position of a field within its table. As this information is important to calculate the role of each Data Vault field, I had to calculate this position in Python.
2. Instead of returning separate columns with `data_type`, `scale`, `precision`, ... the `SHOW` command returns a `JSON` field called `data_type`, with all applicable properties defined there. Example: `{"type":"TIMESTAMP_NTZ","precision":0,"scale":9,"nullable":true}`.
3. For `NUMBER` (or `INTEGER`) fields, `Snowflake` returns `FIXED`, which does not make that much sense, as it is not an actual data type, and cannot even be used in SQL data type conversions (using `::`, or `CAST`). For these fields to be properly instantiated into a `FieldDataType`, we're using `NUMBER` whenever `FIXED` is returned.

**While there:** Added `picnic-guide` to the dependencies and ran `guide lint apply`.