Force CAST to in hash generation when the target field is TEXT (#21)
## Context

In this PR https://github.com/PicnicSupermarket/diepvries/pull/17 we changed the hashing function to ensure every field is casted to its target data type during hash generation. Although, a bug was introduced for String fields.

In the new hashing function, a non-string (an `INTEGER` let's say) uses the following function `COALESCE(CAST(CAST(test_integer AS NUMBER (38, 0)) AS TEXT), '')`. So we initially cast the field to its target data type, then cast it to a string, then `COALESCE` it to an empty string.

For strings, we do not need the double casting, but at least one casting is needed given that the source field may be defined with a different type. At the moment, the framework does this: `COALESCE(test_string, '')` without any casting, when it should do this `COALESCE(CAST(test_string) AS TEXT), '')`.

## Implementation

Applied cast to string before coalescing to empty string.