Add pre-commit tool
In this commit, we add the pre-commit tool to
improve the Software Development Lifecycle
of the project by automatically running checks
(linting & formatting) before one commits their
changes.

Specifically:
- black, flake8 and pylint were adjusted in
tox to run as pre-commit hooks.
Additionally, the check-added-large-files
commit hook was added (prevents giant
files from being committed).

- We fixed doc build and pylint failing:
* Docs were failing with the pinned versions of
their dependencies (sphinx,
sphinx-autodoc-typehints).
So we unpinned the dependencies (also good
as best practice).
* Pylint was not able to read the sections
related to its configuration in tox.ini.
So we added a pylint section in tox.ini with the
necessary key/value entries.
* the snowflake-connector-python was
needed as a dependency for pylint (otherwise
we got an ImportError)

- We updated the CONTRIBUTING.md with
how to install pre-commit